touch_keyboard: Send Key-up when sliding off a key
Previously if a finger pressed down a key and held it, then slid
off the key (without leaving the surface of the sensor) the system
would never send a key-up event, resulting in a constant stream of that
key being "typed" until you pressed another key.  These runaway streams
of keystrokes are obviously very bad for the user.  This CL adds in
a check when a finger moves to make sure that if it has left a key
that is down, a key-up event will be sent.

BUG=b:36232858
TEST=emerged onto my Pbody.  Pressing a key down and sliding off now
stops the stream of keys.  Quick "normal" taps still work fine, and
barely brushing a key still doesn't send events.  Evtest showed the
keyboard events making more sense now too.

Change-Id: Ia4867ea8d66a53e2e84aa2c09ef3677a0666546d
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/455554
Reviewed-by: Jingkui Wang <jkwang@google.com>