touch_keyboard: Reject taps with unusual pressure
Previously pressure values were collected for tapping fingers on the
keyboard, but nothing was really done with it.  Checking pressure values
is a an easy way to reject some bad taps, such as palms and accidentally
brushing the key with a finger.  This CL adds in the ability to check
that the maximum pressure value during a tap is within an acceptable
range.

BUG=b:36279311
TEST=emerged onto my Pbody.  Typing normally on they keyboard still
works, but pressing with my palm doesn't produce key events anymore
nor do very light taps/brushes.

Change-Id: I2b3fb71944a0b3a50b0229a6c65c8881acd02963
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/456044
Reviewed-by: Jingkui Wang <jkwang@google.com>