touch_keyboard: Suppress false error messages
Previously an error message was added to the logs when an event was
about to fire without any current FingerData stored for it unless it's
a guaranteed key-up event.  It turns out that's not the only case where
we'd expect that to happen.  If a finger arrives and leaves all within
the delay window, we might see a guaranteed key-down event too.

This CL edits the error message and the conditions underwhich it is
sent to the logs.  No functional changes here, just trimming out some
extra/misleading error messages.

BUG=b:36275412
TEST=emerged onto my Pbody.  I don't see the error message anymore when
I tap very quickly.  Tapping quickly still sends key events and works
as expected.

Change-Id: I7a7af809dc3f9d7a586c06a07e3ba2936f350f4e
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/456076
Reviewed-by: Jingkui Wang <jkwang@google.com>