imageloader: replace Init function with static factory method
The static factory method prevents use between construction and
initialization by only returning a value if initialization succeeds.
This allows us to get rid of the initialized_ member and several
CHECKs, since users will never see a value if it fails during
LoadManifest. This is preferred style (see go/totw/42).

This also makes IsValidFingerprintFile static since it doesn't
depend on any members, which slightly simplifies a test. The
CopyValidImage test is left untouched because it is marked a friend
and has access to the private constructor and because its use case
is somewhat more complicated.

BUG=None
TEST=unit tests

Change-Id: I1af21db9b3b8af285b191e4982898ddec46e35c9
Reviewed-on: https://chromium-review.googlesource.com/456042
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>