touch_keyboard: fix uinput definition issues
Since the uinputdevice classes use the uinput module, they need to be
compiled for a kernel that includes that module.  The headers for the
kernel must also include the various structs and ioctls as well.
However, we can compile this program using placeholder definitions that
are protected by #ifdef statements to make sure there is a legal
definition at compile time.  Obviously, a copy of this program compiled
this way will only work on a computer with a fully compatible kernel,
but without this patch there's no way to compile against an old kernel.

Essentially this was already happening, but the syscallhandler class
also needed the placeholder definitions for some kernels, so this patch
moves the definitions into their own header that can easily be included
from both syscallhandler.h and uinputdevice.h.  This should allow the
touch_keyboard package to be built for other kernels that don't have
uinput headers availible at compile time (such as will be the case in
the factory branch).

BUG=b:36378443
TEST=built and deploy the package on my machine, make sure everything
compiles and that I can type and use the touchpad on my Pbody.

Change-Id: Ibf9f7cfc540b33dfe5b00e38f92b4ac49f8eaae6
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/457170
Reviewed-by: Marco Chen <marcochen@chromium.org>