authpolicy: Renew user TGT before it expires
Automatically renews a user ticket-granting-ticket (TGT) before it
expires. This prevents that user policy cannot be fetched anymore
if the user session runs for longer than the TGT validity lifetime
(usually 10 hours). The ticket stills expire after the renewal
lifetime (usually 24 hours) or if the machine is offline during all
renewal attempts and logging back in is required. Prompting the user
to relog in Chrome is a future CL. The user TGT will be used for more
services (SAML SSO) in the future.

This CL will become much more important when offline login is
supported. It is then much more likely that the 10 hour limit is
exceeded.

Schedules TGT renewal at 80% of the TGT validity lifetime. If the
renewal fails, another renewal attempt is scheduled at 80% of the
remaining lifetime. This is continued in decreasing intervals no
shorter than 5 minutes to save resources. Automatic renewal is given
up when the TGT expires or the renewal time limit is reached. Renewal
is scheduled on the D-Bus thread on a single-threaded task runner, so
that it won't interfere with D-Bus calls.

BUG=chromium:676607
TEST=Tested on device with a short TGT lifetime of 5 minutes.

Change-Id: Idde26621a72844030aa67903b5d616b6d06a6a62
Reviewed-on: https://chromium-review.googlesource.com/458220
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@google.com>