cryptohome: continue tpm initialization interrupted at srk step
Prior to this change, if tpm initialization was interrupted after
taking the tpm ownership, but before initializing SRK and changing
the owner password, the tpm initialization was never completed.

The reason is: even though InitializeTpm() detects this situation
and continues from the step where it was interrupted, if tpm is
owned, the InitializeTpm() function is never called again.

This situation can be detected by checking the following conditions:
1) the physical tpm is owned;
2) there is no /mnt/stateful_partition/.tpm_owned file, which is
   created only when all initialization steps are done.

This CL makes the following changes:
1) If the incomplete initialization is detected in SetupTpm() as
described above, initiate calling InitializeTpm() to continue
initialization. (Also, update unit tests to check that.)
2) Make sure /mnt/stateful_partition/.tpm_owned is created only
when changing the password from default succeeded (or the default
password doesn't work).

BUG=chromium:652565
TEST=1) Imitate failure after taking ownership and before SRK
     initialization. Check that the initialization is continued
     after restart.
     2) Unit tests (updated).

Change-Id: I48608ffd43838fce419be451aff1135a80be67ea
Reviewed-on: https://chromium-review.googlesource.com/461265
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>