cryptohome: consolidate and fix detecting transient errors
The logic for detecting if the result of a performed action is
a transient error does not depend on tpm version. Move it from
tpm 1.2/2.0-specific code to the common tpm.h

Also, update the logic to include TpmRetryActionRetryReboot into
transient actions - that is one of the flavours of DAD lockouts
for tpm 1.2, and an indication of incorrect TPM reset for tpm 2.0.
Neither of these two (rare, not observed in practice) situations
should be treated as a non-transient error and lead to re-creating
the cryptohome key, for example.

This simplifies the IsTransient logic to: everything but success
and TpmRetryActionFailNoRetry is a transient error.
TpmRetryActionFailNoRetry is what is returned for unknown errors.

BUG=none
TEST=Unit tests.

Change-Id: Iffbb38df7049549658943f411a7db7ca124050cf
Reviewed-on: https://chromium-review.googlesource.com/462001
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>