power: Improve delay scheduling.
Extremely lengthy background information: At present,
powerd's StateController::ScheduleActionTimeout method takes
a bunch of delays for actions (e.g. screen-dim, screen-off)
and a bunch of activities (e.g. user, video, audio) and uses
them to start a "wait until I need to wake up to perform the
next action" timer.

This is all fine and well for activity that's reported
periodically (e.g. video activity, reported every five
seconds), since we'll recompute the time-until-wakeup every
time we get a new report of activity and never actually hit
the timeout.

Things are a bit weirder when it comes to activity that
remains active until we receive notification that it's
stopped. If nothing else is going on, we may actually hit
the timeout after (say) ten minutes. In practice, this is
fine, because the UpdateState method (called by
HandleActionTimeout) recomputes the delays (getting the full
durations again since the activity is still active) and
doesn't do anything.

It makes things a bit weird for tests that want to make sure
that various actions are being blocked, though. Ideally,
they'd be able to just see that the timeout isn't set at
all, but with the way the code works at present, they
actually need to advance the timer until it fires and then
verify that the action wasn't actually performed.

This makes testing much trickier in a followup change, so
I'm introducing new IsIdleBlocked and
IsScreen{Dim,Lock,Off}Blocked methods that can be called to
check whether a given action will be indefinitely blocked
barring a state change. ScheduleActionTimeout now calls
these to verify that a given action can be performed before
computing its delay. If the action is blocked in the current
state, then we don't include its delay when computing the
timeout.

Note also that most actions can't be blocked indefinitely by
anything right now. That will change once Chrome is
reporting wake locks to powerd.

BUG=chromium:704408
TEST=tests pass

Change-Id: Idcbdc36a7fce1784db947b5156013dd2d1a72a9d
Reviewed-on: https://chromium-review.googlesource.com/471831
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>