permission_broker: allow access to ADB
Some Android devices only present an interface for ADB when USB
debugging is enabled, and so are allowed by the unclaimed USB
devices rule. This is the case for e.g. Nexus Player. However,
devices that can also act as mass storage devices often present
several interfaces, most of which will be claimed by usbfs.

permission_broker responds to devices with some claimed and some
unclaimed interfaces by locking down the device, which allows us
access but prevents ADB from operating correctly, so it doesn't
recognize the device as an Android device. This carves out an
exception for this specific type of interface.

BUG=chromium:665622,chromium:703946
TEST=ensure adb running in container can shell into nexus player
  and nexus 7 tablet

Change-Id: Id56f166cf34c178572bfd500b2a2302428b207d0
Reviewed-on: https://chromium-review.googlesource.com/462515
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>