libcontainer: Track device nodes and cgroup perms separately
Tracking the device nodes to create separately from the device node
permissions will allow device permissions to be set for devices that
aren't created by the container startup process.  In new version of the
OCI spec, these two items are in separate areas of config.json and this
change will make migration to that format easier.

The current add_device interface is unchanged and session_manager can
continue to use it when it parses the old 0.2 OCI config for android's
etontainer. run_oci will be changed to use the new interface in a
following commit.

BUG=680859
TEST=updated security_RunOci

Change-Id: I9e2e5932699da8f7406f59677d3a46b946b11664
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/465552