cryptohome: Handle migration of existing links.
If the migration process was stopped after creating a symlink but before
removing the source the process will fail the next time it is run when
trying to re-create the symlink again.  To prevent this we should ignore
EEXIST errors when creating symlinks if the actual target and
destination target are the same.

TEST=cros_workon_make --board=samus --test cryptohome
BUG=b:33710917

Change-Id: I9cfdc97a47c3ecfd9dea1d7924cf058f388c673d
Reviewed-on: https://chromium-review.googlesource.com/465867
Commit-Ready: Dan Spaid <dspaid@chromium.org>
Tested-by: Dan Spaid <dspaid@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>