biod: not storing user id on disk
There are 2 reasons to not store user id on disk: 1. this ensures
that biod code will be compatible during user id migration; 2.
the directory in which records are stored already contain user id
in the form of cryptohome id, so no need to store it to identify
the records.

BUG=chromium:629657
TEST=unittest biod_client_tool

Change-Id: Ib1ae198f236b7530e85f93b8745fd89b23ffcef2
Reviewed-on: https://chromium-review.googlesource.com/466857
Commit-Ready: Mengqi Guo <mqg@chromium.org>
Tested-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>