installer: Use cgpt to bootstrap partition numbers
When we are installing from a payload image, just use cgpt
directly on it to learn what its partition numbers are.

This fixes the case when chromeos-install is run from within
the chroot, because /usr/sbin/write_gpt.sh is not available
inside the chroot.

BUG=chromium:709027
TEST=chromeos-install from chroot

Change-Id: I3eeba7a86e85758259250f46b0d1c3601f0873c7
Reviewed-on: https://chromium-review.googlesource.com/470727
Commit-Ready: Ian Coolidge <icoolidge@google.com>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>