init: Block FIFOs on the stateful file system.
This makes use of the inode security policy code added to the
chromiumos LSM in the kernel. By default, opening of FIFOs on the
stateful file system will now be blocked, as was recently done in
CL:966683 for symlinks. There are exceptions to the FIFO blocking
policy for a couple of file system locations that actually do rely
on FIFOs.

BUG=chromium:655606
TEST=Image still boots and passes tests, no FIFO warnings in dmesg.

Change-Id: I0c54cee75e43b0e8f695da0cd07388038311efbd
Reviewed-on: https://chromium-review.googlesource.com/970905
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>