init: Block symlinks and FIFOs on the stateful file system.
This makes use of the inode security policy code added to the
chromiumos LSM in the kernel. By default, symlink traversal and
opening of FIFOs on the stateful file system are now blocked.
There are exceptions to the symlink policy for a handful of
file system locations that actually do rely on symlinks.

BUG=chromium:655606
TEST=Image still boots and passes tests, no symlink/FIFO warnings
in dmesg.

Change-Id: I9b135c3a4aa4d6534d14d7a4f5c146e0e950d599
Reviewed-on: https://chromium-review.googlesource.com/472908
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>