authpolicy: Add GetUserStatus D-Bus call
Adds a GetUserStatus D-Bus call to be called from Chrome in order to
get the status of a given user account. The status contains account
information, in particular display names, which might change on the
server, as well as the status of the user's Kerberos ticket-granting-
ticket (TGT). Chrome can use this data to prompt the user to relog,
so that a fresh TGT can be acquired (requires password).

CQ-DEPEND=CL:481299

BUG=chromium:710452
TEST=Compiled, ran tests.

Change-Id: I8759ff15dc6aceb57cd3d50953345cb65da4ef25
Reviewed-on: https://chromium-review.googlesource.com/481319
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>