metrics: Add vmlog logging to metrics_daemon.
Historically, we've depended on having direct interaction with systems
to understand swap behavior on systems. This change logs some
information from /proc/vmstat at a high rate so that we get a better
understanding of swap and fault behavior around the time of a feedback
report.

I ran PowerIdleServer to check for any significant power change,
and it was within the noise. w_energy_rate was 6.846 before this
change, and 6.801 after.

BUG=chromium:709662
TEST=unit tests, copied binary and inspected log and rotated output
CL-DEPEND=486090

Change-Id: Idf8767c3b31d11ad09d1929d60f86066d7f8a0d4
Reviewed-on: https://chromium-review.googlesource.com/487085
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>