hal_adapter: employ the ARC++ camera HAL v3 Mojo service model
This CL changes how the Mojo channel is established between various
components in the ARC++ camera HAL v3 stack. A CameraHalDispatcher
Mojo service is started by Chrome and creates the unix domain socket
/var/run/camera/camera3.sock. The CameraHalDispatcher accepts two
types of peers: CameraHalServer and CameraHalClient.

The camera HAL adapter is the CameraHalServer and registers itself
to CameraHalDispatcher; the Chrome VideoCaptureDevicefactory and the
HAL v3 shim in Android cameraserver are the CameraHalClients and
connect to CameraHalDispatcher to ask for camera service. The
CameraHalDispatcher is responsible for relaying the Mojo channel
connection requests from CameraHalClients to the CameraHalServer
and returning the established Mojo channel to the CameraHalClients.

BUG=b:36233258, b:32690003
CQ-DEPEND=CL:544741
TEST=Make sure USB camera preview works in Chrome and Android

Change-Id: I70d623d7ccb316ae9dd261188ccc7c783dab6e00
Reviewed-on: https://chromium-review.googlesource.com/487551
Commit-Ready: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>