debugd: remove unused DBus::Errors
To prepare for replacing dbus-c++ constructs with brillo dbus
bindings, this removes DBus::Error usage which is syntactically
dead (i.e. tools that take one as a parameter but never use it).
This simplifies the interfaces to a number of tools and breaks
the dependency on dbus-c++ constructs entirely for some of them.

BUG=chromium:705051
TEST=emerge, unit tests

Change-Id: Ib3dd26992f6f97173d38008308dad4710456555e
Reviewed-on: https://chromium-review.googlesource.com/486202
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>