debugd: simplify DevFeaturesTool
base::Bind is the preferred method of creating and storing
callbacks, so replace uses of std::bind with that. This also
allows us to more easily use partial application, so we can
pass the DBus::Error pointer it at the callsite to make it
clearer that the error is being set each time a query function
is called.

This also moves some functions which were in the public API but
unused by anything outside of DevFeaturesTool into an anonymous
namespace inside of the .cc file, since they did not touch any
state and were marked const. This allows us to avoid binding
"this" in each of the query functions.

BUG=chromium:705051
TEST=emerge, unit tests

Change-Id: I5c7a333816ea4caa6d90eea0fc028fd7ce3b6d35
Reviewed-on: https://chromium-review.googlesource.com/487071
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>