common: always maps the whole buffer in Lock and LockYCbCr
Calculating the address of a given (x, y) coordinate is error-prone,
especially in multi-planar buffers. To correctly support setting the
coordinate in Lock and LockYCbCr would require a lot of extra codes
and logic inside CameraBufferMapper. Since in the camera world we
normally don't just map a sub-region of the buffer, let's remove
the support for setting the (x, y) coordinate.

The x, y, width, and height args are kept in the Lock and LockYCbCr
APIs to conform with the Android gralloc APIs.

BUG=b:36377554
TEST=unit tests

Change-Id: I0e039cf75da991234848ec84f57477b7df1fa85f
Reviewed-on: https://chromium-review.googlesource.com/489867
Commit-Ready: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Heng-ruey Hsu <henryhsu@google.com>