debugd: rename DebugDaemon to DebugdDBusAdaptor
In a future CL, the DebugDaemon class will contain the DBus adaptor
implementation, and the actual daemon will be a small helper class
in the main file that registers it. This renames the class to better
fit the purpose of this class in the post-dbus-c++ world.

BUG=chromium:705051
TEST=emerge

Change-Id: Ia51e55cc4dc12f6804c514926d0b751479a7f4c7
Reviewed-on: https://chromium-review.googlesource.com/489404
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>