run_oci: remove signature checking
We always use the unsigned option because signature checking is
done when imageloader mounts the squashfs image rather than at
run_oci time. For VMs, this checking happens outside of the VM
so we've already done it by the time run_oci is called in that
case as well. This also removes the need to have libcrypto in the
deps.

BUG=chromium:706615
TEST=emerge, c adb

Change-Id: I7d816df8f69228de80d1c4b4720bf6f1d79bd946
Reviewed-on: https://chromium-review.googlesource.com/494087
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>