debugd: Add Atmel debug logs to chrome://system
Previously when an Atmel touchscreen or touchpad got bad reference
readings the only way for Atmel engineers to get those readings and
tweak their firmware to mitigate the issue they would have to get the
system and reproduce the error themselves with a test image so they can
run their atmel_tools.sh script.  This is posing a difficulty for us
right now for a few ongoing bugs with Atmel touchscreen on
Kevin/Caroline and touchpad on Samus because we have no reliable repro.
It seems that only time & chance factor in.  We have lots of feedback
reports but the Atmel Engineers need these "reference" and "delta"
readings to progress on this issue.

As such, this CL adds in 4 new entries to chrome://system and thereby
the feedback reports sent by users.
  atmel_ts_refs, atmel_ts_deltas, atmel_tp_refs, atmel_tp_deltas

These values pull the "reference" and "delta" readings off any Atmel
touchscreens and touchpads that may be connected to the device.  The
reference readings are essentially baseline capacitance values across
the touch sensor.  The delta values are the difference between the
current capacitance values and those references at the time the command
is run.

As such, these are quite safe data and should not be personally
identifiable or anything like that.  It simply tells the Atmel engineers
a little more about how the touch controller got confused.

BUG=b:37777693
TEST=emerged and deployed debugd onto a Samus (which has both an Atmel
touchscreen and touchpad).  I was able to get the values by querying
debugd over dbus (dbus-send --system --type=method_call --print-reply
--dest=org.chromium.debugd /org/chromium/debugd org.chromium.debugd.GetLog
string:"atmel_ts_refs") for all the 4 new entries.  Similarly, directing
chrome to chrome://system showed all four as well.  On a Snappy (no
Atmel touch devices at all) the entries were empty as expected.
CQ-DEPEND=CL:493829

Change-Id: I42047a295044fcc80d3749c4c2527729533ad7cb
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/493612
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>