chaps: canonicalize XML D-Bus files
The XML specification for the D-Bus interface generally goes in
a dbus_bindings directory for each of our daemons. On top of that,
the file is usually named <service name>.xml. Rename the file and
to fit in with the rest of the daemons and fix the include
directives.

BUG=chromium:719073
TEST=emerge

Change-Id: I73c477108bc380b96080515758dd1f9c174591ee
Reviewed-on: https://chromium-review.googlesource.com/500667
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>