shill: wifi: use CustomAccessor for 'BgscanMethod' property
This CL removes the workaround in WiFi that uses CustomMappedAccessor
for setting the BgscanMethod property because CustomAccessor didn't
provide a way to specify a custom clear method. CL:173321 updated
CustomAccessor to take a custom clear method, so now we can use
CustomAccessor for the BgscanMethod property.

BUG=None
TEST=Run unit tests.

Change-Id: I92576047f16fd1935038a6594a9abb3deafc7679
Reviewed-on: https://chromium-review.googlesource.com/503552
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>