hammerd: Use udev to invoke hammerd.
To simplify the design, we use udev daemon to monitor when the hammer
is connected. This way, hammerd only needs to take care of the logic
of updating the EC firmware.

Note that we need to prevent multiple hammerd processes from executing
at the same time. We implement a file-based lock to accomplish this.

We also add a unittest for the public methods of FirmwareUpdater.

BUG=b:36139110
TEST=FEATURES=test emerge-poppy hammerd
     Connect the hammer and hammerd is invoked by udev
     Manually run hammerd twice and the second one exit immediatedly.
CQ-DEPEND=CL:526873

Change-Id: I7243d0882dcd0bf3f34519c425929122080eea1a
Reviewed-on: https://chromium-review.googlesource.com/520526
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>