metrics: Switch zram metrics to "mm_stat" if it's available
The old way for reading zram metrics was deprecated.  Now we need to
read a file with a bunch of space-separated values.

The kernel describes this file as:

  File /sys/block/zram<id>/mm_stat

  The stat file represents device's mm statistics. It consists of a
  single line of text and contains the following stats separated by
  whitespace:
     orig_data_size
     compr_data_size
     mem_used_total
     mem_limit
     mem_used_max
     zero_pages
     num_migrated

For a while now on kernels like 4.4 we've seen a warning in dmesg at
every bootup about using the old deprecated attribute.  On newer
upstream kernels the deprecated attribute is totally gone.

NOTE that on older kernels (like 3.14) the "mm_stat" file doesn't
exist, so we need to keep support for the old location too.

BUG=chromium:721955
TEST=cros_run_unit_tests --board ${BOARD} --packages "metrics"
TEST=Put printouts and see that metrics got read OK
TEST=Boot up and no more warning in the kernel logs

Change-Id: I46478dc59f6c875b3b93db0e7e7524002b0f1496
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/506370
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>