chaps: add D-Bus constants
Because of chaps-on-linux, we can't use chromeos-dbus-bindings to
generate the D-Bus adaptors and proxies for chaps, and we're going
to need to use the libchrome bindings and libbrillo helpers
directly. This means we need to put the strings for method names
somewhere. Luckily, the proxy and adaptor implementations both
live in this repo so we don't need to export it in system_api, and
we can just use them internally in chaps.

BUG=chromium:719073
TEST=emerge

Change-Id: Ie804d539fddfd75504e334e164a72eb0fb2ed7b4
Reviewed-on: https://chromium-review.googlesource.com/508254
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>