common-mk: Create common platform2 testrunner lib
Creates a common testrunner lib under common-mk in an effort to
consolidate the over 25 copies of testrunner currently spread
throughout platform2.

As a simple demonstration, I migrated mist to use this common lib.
Once this patch set is agreed upon, I'll migrate the remaining
implementations to use the common library.

If there are conflicts that can't be resolved with the testrunner.cc
main implementation (during the migration), I'll extend this pattern
as necessary with customized libraries for each unique scenario.

BUG=chromium:456569
TEST=unit test refactor

Change-Id: I9e9a173deac2cb84ce94baf9821c22d469b978b7
Reviewed-on: https://chromium-review.googlesource.com/508239
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Dan Erat <derat@chromium.org>