chaps: simplify initialization path
ChapsServiceImpl::GetSlotList only relies on bits of initialization
in SlotManagerImpl that can't fail, so we can move those into the
constructor to guarantee they will be done by the time the
ChapsServiceImpl is constructed. Because of this we no longer need
to do any special initialization/teardown for ChapsServiceImpl.

BUG=None
TEST=chaps/cryptohome unit tests, autotests:
  platform_Pkcs11Events platform_Pkcs11LoadPerf
  platform_Pkcs11InitOnLogin platform_Pkcs11ChangeAuthData
  platform_TPMEvict platform_Pkcs11InitUnderErrors

Change-Id: I9efc18d5d5d0e4c23f61cd67f8fed043d8c7282e
Reviewed-on: https://chromium-review.googlesource.com/508875
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>