debugd: don't close bound fds in SandboxedProcess
dbus::FileDescriptor contains a base::ScopedFD, which fails a
CHECK on destruction if the close call fails with EBADF. This
means you can't double-close a ScopedFD. This is a problem if
it's stored in a SandboxedProcess that doesn't outlive the
dbus::FileDescriptor because the ProcessImpl closes all FDs in
its pipe map on destruction.

Unfortunately, the dbus::FileDescriptors are passed as const &
from the D-Bus bindings, so TakeValue() is not an option
without a const cast. Additionally, if BindFd was supposed to
take ownership of the FD its API should take a ScopedFD instead
to signal this.

BUG=chromium:722501
TEST=use SystraceTool, PerfTool and ensure they don't crash
  debugd via the double-close CHECK; also watch the FD table
  and make sure no FDs are leaked

Change-Id: I0b748e34f311f04dda33287f4e85ce4bf6970e27
Reviewed-on: https://chromium-review.googlesource.com/514522
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>