hal_adapter: use InterfaceRequest in OpenDevice call
Instead of asking a remote Mojo implementation to return an InterfacePtr,
the preferred way to set up a Mojo connection to a remote Mojo implementation
is to send a InterfaceRequest. The InterfaceRequest includes a message pipe
handle which the remote Mojo implementation can bind to its Mojo binding.

BUG=b:32690003
TEST=Make sure camera works in Chrome and Android container.

Change-Id: Ic0d905c8465f0a0fda6b18ec2b06b9352dc3bdba
Reviewed-on: https://chromium-review.googlesource.com/527849
Commit-Ready: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Heng-ruey Hsu <henryhsu@google.com>