chaps: rearrange initialization code
The TPM2 base::Thread requires a base::AtExitManager, but we don't
create one of those until construction of the DBusServiceDaemon,
which comes later. Instantiating an AtExitManager before that will
result in a DCHECK failure since AtExitManager is a singleton and
we would be trying to make two.

In this case, we instantiate the Daemon first, and then do the
formerly main-function initialization inside of the OnInit method.
This ensures an AtExitManager exists before trying to instantiate
any threads.

Additionally, the async init thread is changed to a base::Thread
from a PlatformThread to reduce some boilerplate and also because
the base::Thread destructor will ensure the thread is stopped
before we destroy the Daemon.

BUG=chromium:727987
TEST=deploy to cyan device and run (since thread construction is
  no longer behind USE_TPM2), use pkcs11-tool, build on reef

Change-Id: I24584d640ea509cb19aa0358a3f7e304560d0a09
Reviewed-on: https://chromium-review.googlesource.com/520163
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>